Information on debounce routines
Based on hardware and gameboy ROMs, there don't appear to be pullup/pulldown resistors for debouncing. In most cases, a sophisticated debounce routine will not be required by the game genres common to the Game Boy ecosystem. However, investigation of Super Mario Land shows a two stage debounce routine.

First, multiple reads generate the delay necessary to bleed out capacitance built up in the joypad for a clear and stable reading. Then, `XOR` and `AND` compare a new reading with a previous one, helping to significantly reduce the presence of false readings.

Many emulators will not have artificial bounce (it's often filtered out at OS and driver level), so it is necessary to test debounce on hardware.